+++ OUDS/Core/Tokens/SemanticTokens/Sources/ElevationSemanticTokens.swift
@@ -22,11 +22,13 @@ public typealias ElevationYSemanticToken = ElevationRawToken
 public typealias ElevationBlurSemanticToken = ElevationRawToken
 public typealias ElevationSpreadSemanticToken = ElevationRawToken
 public typealias ElevationColorSemanticToken = ColorRawToken
+public typealias ElevationBoxShadowSemanticToken = ElevationBoxShadowRawToken
 
 // MARK: - Semantic tokens
 
 /// This is a group of semantic tokens for **elevations**.
-/// It defines all `ElevationSemanticToken` a theme must have.
+/// It defines all elevation semantic tokens a theme must have. (`ElevationZIndexSemanticToken`,`ElevationXSemanticToken`,`ElevationYSemanticToken`,
+/// `ElevationBlurSemanticToken`,`ElevationSpreadSemanticToken` and `ElevationColorSemanticToken`)
 public protocol ElevationSemanticTokens {
 
     // MARK: Semantic token - Elevation - Z index
@@ -56,7 +58,7 @@ public protocol ElevationSemanticTokens {
     var elevationXStickyDefault: ElevationZIndexSemanticToken { get }
     var elevationXStickyEmphasis: ElevationZIndexSemanticToken { get }
     var elevationXStickyNavigationScrolled: ElevationZIndexSemanticToken { get }
-    var elevationXFOcus: ElevationZIndexSemanticToken { get }
+    var elevationXFocus: ElevationZIndexSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Y
 
@@ -96,55 +98,74 @@ public protocol ElevationSemanticTokens {
 
     // MARK: Semantic token - Elevation - Color - None
 
-    var elevationColorNoneLigh: ElevationColorSemanticToken { get }
+    var elevationColorNoneLight: ElevationColorSemanticToken { get }
     var elevationColorNoneDark: ElevationColorSemanticToken { get }
-    var elevationColorNoneInverse: ElevationColorSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Color - Raised
 
-    var elevationColorRaisedLigh: ElevationColorSemanticToken { get }
+    var elevationColorRaisedLight: ElevationColorSemanticToken { get }
     var elevationColorRaisedDark: ElevationColorSemanticToken { get }
-    var elevationColorRaisedInverse: ElevationColorSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Color - Drag
 
-    var elevationColorDragLigh: ElevationColorSemanticToken { get }
+    var elevationColorDragLight: ElevationColorSemanticToken { get }
     var elevationColorDragDark: ElevationColorSemanticToken { get }
-    var elevationColorDragInverse: ElevationColorSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Color - Overlay - Default
 
-    var elevationColorOverlayDefaultLigh: ElevationColorSemanticToken { get }
+    var elevationColorOverlayDefaultLight: ElevationColorSemanticToken { get }
     var elevationColorOverlayDefaultDark: ElevationColorSemanticToken { get }
-    var elevationColorOverlayDefaultInverse: ElevationColorSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Color - Overlay - Emphasis
 
-    var elevationColorOverlayEmphasisLigh: ElevationColorSemanticToken { get }
+    var elevationColorOverlayEmphasisLight: ElevationColorSemanticToken { get }
     var elevationColorOverlayEmphasisDark: ElevationColorSemanticToken { get }
-    var elevationColorOverlayEmphasisInverse: ElevationColorSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Color - Sticky - Default
 
-    var elevationColorStickyDefaultLigh: ElevationColorSemanticToken { get }
+    var elevationColorStickyDefaultLight: ElevationColorSemanticToken { get }
     var elevationColorStickyDefaultDark: ElevationColorSemanticToken { get }
-    var elevationColorStickyDefaultInverse: ElevationColorSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Color - Sticky - Emphasis
 
-    var elevationColorStickyEmphasisLigh: ElevationColorSemanticToken { get }
+    var elevationColorStickyEmphasisLight: ElevationColorSemanticToken { get }
     var elevationColorStickyEmphasisDark: ElevationColorSemanticToken { get }
-    var elevationColorStickyEmphasisInverse: ElevationColorSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Color - Sticky - Navigation scrolled
 
-    var elevationColorStickyNavigationScrolledLigh: ElevationColorSemanticToken { get }
+    var elevationColorStickyNavigationScrolledLight: ElevationColorSemanticToken { get }
     var elevationColorStickyNavigationScrolledDark: ElevationColorSemanticToken { get }
-    var elevationColorStickyNavigationScrolledInverse: ElevationColorSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Color - Focus
 
-    var elevationColorFocusLigh: ElevationColorSemanticToken { get }
+    var elevationColorFocusLight: ElevationColorSemanticToken { get }
     var elevationColorFocusDark: ElevationColorSemanticToken { get }
-    var elevationColorFocusInverse: ElevationColorSemanticToken { get }
+
+    // MARK: Semantic token - Elevation - Bow shadow
+
+    // TODO: Missing elevationBoxShadowNoneLight semantic token
+
+    var elevationBoxShadowRaisedLight: ElevationBoxShadowSemanticToken { get }
+    var elevationBoxShadowRaisedDark: ElevationBoxShadowSemanticToken { get }
+
+    var elevationBoxShadowDragLight: ElevationBoxShadowSemanticToken { get }
+    var elevationBoxShadowDragDark: ElevationBoxShadowSemanticToken { get }
+
+    var elevationBoxShadowOverlayDefaultLight: ElevationBoxShadowSemanticToken { get }
+    var elevationBoxShadowOverlayDefaultDark: ElevationBoxShadowSemanticToken { get }
+
+    var elevationBoxShadowOverlayEmphasisLight: ElevationBoxShadowSemanticToken { get }
+    var elevationBoxShadowOverlayEmphasisDark: ElevationBoxShadowSemanticToken { get }
+
+    var elevationBoxShadowStickyDefaultLight: ElevationBoxShadowSemanticToken { get }
+    var elevationBoxShadowStickyDefaultDark: ElevationBoxShadowSemanticToken { get }
+
+    var elevationBoxShadowStickyEmphasisLight: ElevationBoxShadowSemanticToken { get }
+    var elevationBoxShadowStickyEmphasisDark: ElevationBoxShadowSemanticToken { get }
+
+    var elevationBoxShadowStickyNavigationScrolledLight: ElevationBoxShadowSemanticToken { get }
+    var elevationBoxShadowStickyNavigationScrolledDark: ElevationBoxShadowSemanticToken { get }
+
+    var elevationBoxShadowFocusLight: ElevationBoxShadowSemanticToken { get }
+    var elevationBoxShadowFocusDark: ElevationBoxShadowSemanticToken { get }
 }